completed task

+++ frontend/src/App.js
@@ -1,5 +1,8 @@
 import {useState, useEffect} from 'react';
 import { getTasks, createTask, updateTask, deleteTask, toggleTask } from "./services/api";
+import "./styles/App.css";
+
+import TaskFilter from './components/TaskFilter';
 
 import TaskCarousel from './components/TaskCarousel';
 import TaskForm from './components/TaskForm';
@@ -9,6 +12,7 @@ function App() {
   const [tasks, setTasks] = useState([]);
   const [loading, setLoading] = useState(true);
   const [error, setError] = useState(null);
+  const [filter, setFilter] = useState('all');
 
   const [editingTask, setEditingTask] = useState(null);
 
@@ -20,6 +24,17 @@ function App() {
     }
     setEditingTask(null);
   }
+
+
+  const filteredTasks = tasks.filter(task => {
+    if (filter === 'completed'){
+        return task.completed
+    }
+    if(filter === 'pending') {
+        return !task.completed
+    }
+    return true
+  })
 
 
 async function createHandler(task) {
@@ -43,8 +58,8 @@ async function ToggleHandler(taskId) {
 }
 
 async function deleteHandler(taskId) {
+    if (!window.confirm('Delete this task?')) return;
     try {
-
         await deleteTask(taskId);
         setTasks(tasks.filter(task => task.id !== taskId));
         setError(null);
@@ -80,15 +95,20 @@ useEffect(() => {
 }, []);
 
 return (
-    <div>
+    <div className='app'>
         <h1>Task List</h1>
         {loading && <p>Loading...</p>}
         {error && <p>Error: {error}</p>}
-        <TaskCarousel tasks={tasks}
+       {!loading && 
+       <>
+        <TaskFilter filter={filter} onFilterChange={setFilter} />
+        <TaskCarousel tasks={filteredTasks}
          onToggle={ToggleHandler}
         onDelete={deleteHandler}
         onEdit={setEditingTask} />
         <TaskForm AddTask={submitHandler} editingTask={editingTask} />
+       </>
+       }
     </div>
 );
